Add files via upload
This new version of the protected mode code is able to read a sector from a floppy diskette.  It bypasses the issue that the first application had with determining the proper DMA buffer address for the disk controller by disabling DMA mode.  It just reads data bytes from the disk controller data register.